Hang onto and expand the same "next expansion focus" point
This is due to a limitation in how GeoJS works: the mouse coordinates
are not properly re-read on mousewheel events (but they are on
mousemove), so the workaround is to allow mousemove to clear the focus
point, but maintain it otherwise, always pushing the expansion factor
into it during a sequence of non-moving mousewheel events.

@manthey advised me on this problem, and I've opened [a GeoJS
issue](https://github.com/OpenGeoscience/geojs/issues/1025) about it.